+++ src/main/java/com/projectsassy/sassy/token/TokenProvider.java
@@ -29,7 +29,6 @@ public class TokenProvider {
     // JWT 토큰에 관련된 암호화, 복호화, 검증 로직
 
     private static final String AUTHORITIES_KEY = "auth";
-    private static final String BEARER_TYPE = "Bearer";
     private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 30; //30분
     private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 60 * 24 * 7; // 7일
